requirements.txt
Django
djangorestframework
djangorestframework_simplejwt
django-cors-headers
channels[daphne]
channels_redis
Pillow
asgiref
python-dotenv
redis